Arrow backend (#13)
* Add columnar metadata boundary

* Add Arrow metadata bridge

* Add Arrow-buffered VCF metadata path

* Extend Arrow-backed VCF genotype metadata

* Refactor VCF text metadata dispatch

* Normalize Python metadata boundary to Arrow

* Route text VCF row APIs through Arrow metadata

* Move PyO3 matrix reads to Arrow IO facades

* Make PLINK1 reads Arrow-backed

* Make PLINK2 hardcall reads Arrow-backed

* Make PLINK2 dosage and haplotype reads Arrow-backed

* Make BGEN dense reads Arrow-backed

* Make BCF reads Arrow-backed

* Remove row metadata adapters from IO reads

* Remove legacy row metadata adapters

* perf: avoid sparse index bridge copy

* Make sample metadata columnar

* Add borrowed variant metadata views

* Optimize VCF text metadata reads

* Document VCF text metadata fast paths

* Reuse BCF decode buffers

* Stream PLINK metadata reads

* Avoid PGEN difflist entry allocations

* Optimize BGEN metadata skips

* Document BGEN scratch metadata reads

* Tighten dense dtype assembly

* Rename Rust output API

+++ tests/test_rust_numpy_conversion.py
@@ -158,8 +158,8 @@ def test_rust_sparse_indices_transfer_ownership_to_numpy_without_bytearray_base(
 
     assert isinstance(indptr, np.ndarray)
     assert isinstance(indices, np.ndarray)
-    assert indptr.dtype == np.dtype("int64")
-    assert indices.dtype == np.dtype("int64")
+    assert indptr.dtype == np.dtype("int32")
+    assert indices.dtype == np.dtype("int32")
     assert not _is_bytearray_backed(indptr)
     assert not _is_bytearray_backed(indices)
